requirements.txt
langchain
langchain-google-genai
langchain_openai
langchain_anthropic
langchain_community
langgraph
langsmith
streamlit
python-dotenv
agentic-doc
pdfplumber
chromadb
unstructured[pptx]
docx2txt
langchain-chroma
langchain-mcp-adapters
langgraph-cli



